refactor: adding a new dto for update players, and using it on update ep

+++ src/jogadores/jogadores.service.ts
@@ -3,6 +3,7 @@ import { CriarJogadorDto } from './dto/criar-jogador.dto';
 import { Jogador } from './interfaces';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
+import { AtualizarJogadorDto } from './dto/atualizar-jogador.dto';
 
 @Injectable()
 export class JogadoresService {
@@ -11,7 +12,7 @@ export class JogadoresService {
     private readonly jogadorModel: Model<Jogador>,
   ) {}
 
-  async criarJogador(criarJogadorDto: CriarJogadorDto): Promise<void> {
+  async criarJogador(criarJogadorDto: CriarJogadorDto): Promise<Jogador> {
     const { email } = criarJogadorDto;
 
     const jogadorEncontrado = await this.jogadorModel.findOne({ email }).exec();
@@ -20,15 +21,17 @@ export class JogadoresService {
       throw new NotFoundException(`Jogador com e-mail ${email} já cadastrado`);
     }
 
-    await this.jogadorModel.create(criarJogadorDto);
+    return this.jogadorModel.create(criarJogadorDto);
   }
 
   async atualizarJogador(
     _id: string,
-    criarJogadorDto: CriarJogadorDto,
-  ): Promise<void> {
-    await this.jogadorModel
-      .findOneAndUpdate({ _id }, { $set: criarJogadorDto })
+    atualizarJogador: AtualizarJogadorDto,
+  ): Promise<Jogador> {
+    console.log(atualizarJogador);
+
+    return this.jogadorModel
+      .findOneAndUpdate({ _id }, { $set: atualizarJogador })
       .exec();
   }
 
@@ -40,13 +43,19 @@ export class JogadoresService {
     const jogadorEncontrado = await this.jogadorModel.findOne({ _id }).exec();
 
     if (!jogadorEncontrado) {
-      throw new NotFoundException(`Jogador com e-mail ${_id} não encontrado`);
+      throw new NotFoundException(`Jogador com id ${_id} não encontrado`);
     }
 
     return jogadorEncontrado;
   }
 
   async deletarJogador(_id: string): Promise<void> {
+    const jogadorEncontrado = await this.jogadorModel.findOne({ _id }).exec();
+
+    if (!jogadorEncontrado) {
+      throw new NotFoundException(`Jogador com id ${_id} não encontrado`);
+    }
+
     await this.jogadorModel.deleteOne({ _id }).exec();
   }
 }
